updated member size of FS_Info struct to be type long long to accomodate larger file sizes

+++ src/fs_info.h
@@ -8,7 +8,7 @@
 */
 typedef struct FS_Info {
     char name[1024];
-    int size;
+    long long size;
 } FS_Info ;
 
 #endif
